Milestone - BDT code working
I reverted the master branch back to Danny's code before I started messing with things. I needed to tweak it a little bit (e.g., changing the ID number for the NBA, squashing a few knowns bugs that he'd fixed a few days ago), but it's pretty close to what he had. That code mostly worked in production, but it was missing some functionality (e.g., time of the read/write calls to Qualtrics embedded data) and didn't reflect a few changes we'd since made to the survey itself.

The code version pushed here is working, I think. Now I need to edit it to make the survey look & work as we want for launch.

I'm actively updating this code. I'm just merging these branches as this is a nice milestone.